+++ unity_app/Assets/Scripts/DropdownUtils/DropdownIceServersProtocol.cs
@@ -1,0 +1,12 @@
+﻿/*
+ * Copyright 2022 RICOH Company, Ltd. All rights reserved.
+ */
+using com.ricoh.livestreaming;
+
+public class DropdownIceServersProtocol : DropdownEnumBase<IceServersProtocol>
+{
+    override protected string GetItemName(IceServersProtocol item)
+    {
+        return item.ToString().ToUpper();
+    }
+}