+++ unity_app/Assets/Scripts/DropdownUtils/DropdownMuteType.cs
@@ -1,29 +1,8 @@
-﻿using com.ricoh.livestreaming;
-using System;
+﻿/*
+ * Copyright 2022 RICOH Company, Ltd. All rights reserved.
+ */
+using com.ricoh.livestreaming;
 
-public class DropdownMuteType : DropdownInheritance
+public class DropdownMuteType : DropdownEnumBase<MuteType>
 {
-    internal MuteType MuteType { get; private set; } = MuteType.Unmute;
-    private Action<MuteType> onValueChangedAction;
-
-    internal void Initialize(Action<MuteType> onValueChangedAction = null)
-    {
-        Initialize(true);
-        this.onValueChangedAction = onValueChangedAction;
-    }
-
-    override internal void OnValueChangedInternal()
-    {
-        MuteType = (MuteType)Enum.ToObject(typeof(MuteType), value);
-        onValueChangedAction?.Invoke(MuteType);
-    }
-
-    override internal void Refresh()
-    {
-        ClearOptions();
-        foreach (var type in Enum.GetValues(typeof(MuteType)))
-        {
-            options.Add(new OptionData(type.ToString()));
-        }
-    }
 }