+++ unity_app/Assets/Scripts/RoomSpec.cs
@@ -1,26 +1,27 @@
-﻿using System;
+﻿using Newtonsoft.Json;
+using Newtonsoft.Json.Converters;
 using System.Collections.Generic;
+using System.Runtime.Serialization;
 
 public class RoomSpec
 {
+    [JsonConverter(typeof(StringEnumConverter))]
     public enum Type
     {
+        [EnumMember(Value = "sfu")]
         Sfu,
+        [EnumMember(Value = "sfu_large")]
+        SfuLarge,
+        [EnumMember(Value = "p2p")]
         P2p,
+        [EnumMember(Value = "p2p_turn")]
         P2pTurn
     }
 
     private readonly Type type;
 
-    private readonly string[] typeStrings = { "sfu", "p2p", "p2p_turn" };
-
     public RoomSpec(Type type)
     {
-        if (!Enum.IsDefined(typeof(Type), type))
-        {
-            throw new ArgumentOutOfRangeException(nameof(type));
-        }
-
         this.type = type;
     }
 
@@ -28,8 +29,8 @@ public class RoomSpec
     {
         var dic = new Dictionary<string, object>
         {
-            ["type"] = typeStrings[(int)type],
-//            ["media_control"] = new Dictionary<string, object>() { ["bitrate_reservation_mbps"] = 10 }
+            ["type"] = type,
+            ["media_control"] = new Dictionary<string, object>() { ["bitrate_reservation_mbps"] = 25 }
         };
 
         return dic;